+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @orquestra/core
 
+## 0.1.0
+
+### Minor Changes
+
+- add provision and deprovision methods to bootstrap manager and Orquestra
+
 ## 0.0.2
 
 ### Patch Changes